requirements.txt
flask
flask_restful
spacy-nightly[transformers]

--find-links https://download.pytorch.org/whl/torch_stable.html

torch==1.7.0+cpu
torchvision==0.8.1+cpu
torchaudio===0.7.0
